% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPlotGrids.r
\name{addGrid}
\alias{addGrid}
\title{add grid(s) to xyplots
Treats date axis and log10 axis as special cases
For log10 axis:
  Major grids are placed at integral powers of 10 inside the data range
  Minor grids are placed at integral multiples of each power of 10, e.g. 1,2,...9,10,20, 30, ....
  Minor grids may extend past end of data
For date axis
  Grids added at each of the "at" points for each scale.
  This requires that the orignal plot be created with explicit data ticks, e.g.
  scales = list(x=list(at=as.Date(pretty_dates(df$date, 5)), rot=45))}
\usage{
addGrid(p, MinorLines = TRUE)
}
\description{
Always draws major grid lines
Draws minor grid lines if MinorLines is TRUE
}
