% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlot.r
\name{covidPlot}
\alias{covidPlot}
\title{Plot covid data in my format
Parameters are same as for xyplot, except that
   * I have not yet implemented subset as an argument
   * logX, logY determine if the axis are log or not (FALSE = linear, True = Log10, N = LogN)
   * formatX, formatY are the format to apply to axis ticks
   * numTickIntervalsX, numTickIntervalsY are the approximate number of intervals between tick marks}
\usage{
covidPlot(
  formula1,
  data,
  groups = NULL,
  logX = FALSE,
  logY = TRUE,
  type = c("p", "l"),
  subtitle = "",
  xlab = NULL,
  numTickIntervalsX = 5,
  formatX = NULL,
  ylab = NULL,
  numTickIntervalsY = 5,
  formatY = NULL,
  allow.multiple = is.null(groups) || outer,
  outer = !is.null(groups),
  drop.unused.levels = lattice.getOption("drop.unused.levels"),
  ...
)
}
\description{
Plot covid data in my format
Parameters are same as for xyplot, except that
   * I have not yet implemented subset as an argument
   * logX, logY determine if the axis are log or not (FALSE = linear, True = Log10, N = LogN)
   * formatX, formatY are the format to apply to axis ticks
   * numTickIntervalsX, numTickIntervalsY are the approximate number of intervals between tick marks
}
