% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.r
\name{daily_ticks}
\alias{daily_ticks}
\title{Calculate daily tick marks
x, the data range to consider
numIntervals, the approximate number of intervals between tick marks}
\usage{
daily_ticks(x, numIntervals = 3)
}
\description{
Calculate daily tick marks
x, the data range to consider
numIntervals, the approximate number of intervals between tick marks
}
