% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTicks.r
\name{date_ticks}
\alias{date_ticks}
\title{Calculate tick marks for a date axis
x, the data range to consider
dayOfWeek, the day to tick on (0 = sunday) (only applies if weekly ticks)
numIntervals, the approximate number of intervals between tick marks}
\usage{
date_ticks(x, numIntervals = 3, weekStartDay = 0)
}
\description{
Calculate tick marks for a date axis
x, the data range to consider
dayOfWeek, the day to tick on (0 = sunday) (only applies if weekly ticks)
numIntervals, the approximate number of intervals between tick marks
}
