% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.r
\name{linear_ticks}
\alias{linear_ticks}
\title{return a linear sequence including all data in x with approximately numIntervals}
\usage{
linear_ticks(x, numIntervals = 3)
}
\description{
return a linear sequence including all data in x with approximately numIntervals
}
