% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPlotGrids.r
\name{log_ticks}
\alias{log_ticks}
\title{Calculate tick locations that will work nicely with log 10 grids
returns a list of major tick locations and minor tick locations,
$majors and $minors}
\usage{
log_ticks(log_range)
}
\description{
Calculate tick locations that will work nicely with log 10 grids
returns a list of major tick locations and minor tick locations,
$majors and $minors
}
