% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTicks.r
\name{monthly_ticks}
\alias{monthly_ticks}
\title{Calculate monthly tick marks
x, the data range to consider
numIntervals, the approximate number of intervals between tick marks}
\usage{
monthly_ticks(x, numIntervals = 3)
}
\description{
Calculate monthly tick marks
x, the data range to consider
numIntervals, the approximate number of intervals between tick marks
}
