% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTicks.r
\name{nice_step_size}
\alias{nice_step_size}
\title{Get step sizes that are  a nice multiple of 1, 2, 5, or 10
range is max - min
numSteps is the approximate number of steps desired in range}
\usage{
nice_step_size(range, numSteps)
}
\description{
Get step sizes that are  a nice multiple of 1, 2, 5, or 10
range is max - min
numSteps is the approximate number of steps desired in range
}
