% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresidentPlot.r
\name{PresidentPlot}
\alias{PresidentPlot}
\title{xyplot with presidential party shown by background}
\usage{
PresidentPlot(formula, data, inagurations = NULL, ...)
}
\description{
xyplot with presidential party shown by background
}
