% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Presidents.r
\name{Presidents}
\alias{Presidents}
\title{a simple S4 class containing presidential inaguration dates, party, party color, and a few getters}
\usage{
Presidents()
}
\description{
a simple S4 class containing presidential inaguration dates, party, party color, and a few getters
}
