% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPlot.r
\name{symmetricPlot}
\alias{symmetricPlot}
\title{Create a "symmetric log plot", with postive values on a log plot going up from zero, zero values at 0,
 and negitive values as a plot of log of absolute error going down from zero
 Parameters are same as for xyplot, except that
   * I have not yet implemented subset as an argument
   * logX, logY determine if the axis are log or not (FALSE = linear, True = Log10, N = LogN)
   * formatX, formatY are the format to apply to axis ticks
   * numTickIntervalsX, numTickIntervalsY are the approximate number of intervals between tick marks}
\usage{
symmetricPlot(
  formula1,
  data,
  groups = NULL,
  subtitle = "",
  numTickIntervals = 5,
  xlab = NULL,
  numTickIntervalsX = 5,
  formatX = NULL,
  ylab = NULL,
  formatY = NULL,
  type = c("p", "l"),
  allow.multiple = is.null(groups) || outer,
  outer = !is.null(groups),
  drop.unused.levels = lattice.getOption("drop.unused.levels"),
  ...
)
}
\description{
Create a "symmetric log plot", with postive values on a log plot going up from zero, zero values at 0,
 and negitive values as a plot of log of absolute error going down from zero
 Parameters are same as for xyplot, except that
   * I have not yet implemented subset as an argument
   * logX, logY determine if the axis are log or not (FALSE = linear, True = Log10, N = LogN)
   * formatX, formatY are the format to apply to axis ticks
   * numTickIntervalsX, numTickIntervalsY are the approximate number of intervals between tick marks
}
