% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.r
\name{weekly_ticks}
\alias{weekly_ticks}
\title{Calculate weekly tick marks
x, the data range to consider
dayOfWeek, the day to tick on (0 = sunday)
numIntervals, the approximate number of intervals between tick marks}
\usage{
weekly_ticks(x, numIntervals = 3, dayOfWeek = 0)
}
\description{
Calculate weekly tick marks
x, the data range to consider
dayOfWeek, the day to tick on (0 = sunday)
numIntervals, the approximate number of intervals between tick marks
}
