% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks.r
\name{log_ticks}
\alias{log_ticks}
\title{Calcualte ticks for one axis of a log plot
x is the data to fit, in linear coordinates
base is the log base to use, typically 2 or 10 or powers therof
returns a list of
   ticksAt (tick location in original coordinates),
   majors (major grid lines in log coordinates)
   minors (minor grid lines in log coordinates)}
\usage{
log_ticks(x, base = 10)
}
\description{
minors will be null if base is not a power of 10
}
